chore: disable Prettier format-on-save to reduce noise
