+++ src/components/customer/ui/index.ts
@@ -10,4 +10,3 @@ export * from "./calendar-popover";
 export * from "./calendar";
 export * from "./notification-container";
 export * from "./notification-popover";
-export * from "./calendar-popover-new";
