+++ src/utils/misc.ts
@@ -55,14 +55,3 @@ export const formatPrice = (price: number) => {
     maximumFractionDigits: 0,
   }).format(price);
 };
-
-// const formatCurrency = (amount: string | number) => {
-//   return new Intl.NumberFormat("en-NG", {
-//     style: "currency",
-//     currency: "NGN",
-//     minimumFractionDigits: 0,
-//     maximumFractionDigits: 0,
-//   })
-//     .format(Number(amount))
-//     .replace("NGN", "₦");
-// };
